requirements.txt
streamlit
python-dotenv
spacy
https://github.com/explosion/spacy-models/releases/download/en_core_web_trf-3.7.2/en_core_web_trf-3.7.2.tar.gz
langchain
openai
pandas 
sentence-transformers
umap-learn
scikit-learn
hdbscan
plotly
tqdm